refactor: move module descriptors to Vera paths
