+++ proto/vera/bulletin/module/module.proto
