+++ proto/vera/core/module/module.proto
