+++ x/bulletin/module/module.go
@@ -18,7 +18,7 @@ import (
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 
-	modulev1 "github.com/sourcenetwork/sourcehub/api/sourcehub/bulletin/module"
+	modulev1 "github.com/sourcenetwork/sourcehub/api/vera/bulletin/module"
 	"github.com/sourcenetwork/sourcehub/app/metrics"
 	acpkeeper "github.com/sourcenetwork/sourcehub/x/acp/keeper"
 	"github.com/sourcenetwork/sourcehub/x/bulletin/keeper"
